Split browser.py into modules; significant speed optimizations for wd.py
