+++ isoplots/isonice/browser.py
@@ -4,36 +4,27 @@ Usage:
     2. python isonice/isofit.py
     3. Acknowledge this is a WORK IN PROGRESS
 """
-import configparser
+
 import json
 import logging
 import re
-import socket
-from asyncio import create_task, sleep
+
+from asyncio import create_task
 from pathlib import Path
-from types import SimpleNamespace
 
 import click
-import plotly.express as px
-import plotly.graph_objects as go
-import xarray as xr
+from nicegui import (
+    app,
+    run,
+    ui
+)
 
-from nicegui import app, observables, run, ui
+from isoplots.isonice import Config
+from isoplots.isonice.tabs import Tabs
 
-from isofit.radiative_transfer import luts
 
-Config = configparser.ConfigParser()
-Logger = logging.getLogger("IsoNice")
+Logger = logging.getLogger("App")
 
-# TODO: This will source from isofit once the PR is accepted
-try:
-    from isofit.utils.wd import IsofitWD, Loaders
-except:
-    Logger.exception("Failed to load IsofitWD from ISOFIT, using fallback")
-    from isoplots.isonice.wd import IsofitWD, Loaders
-
-# Global shared working directory
-WD = IsofitWD(".")
 
 dark = ui.dark_mode()
 dark.enable()
@@ -170,1835 +161,8 @@ ui.add_head_html('''
 ''')
 
 
-def plotlyColor(i):
-    """
-    Retrieves a plotly color
-
-    Parameters
-    ----------
-    i : int
-        Index color, wraps if the index exceeds to length of the color list
-
-    Returns
-    -------
-    str
-        Hex color
-    """
-    c = px.colors.qualitative.Plotly
-    return c[i % len(c)]
-
-
-class Logs:
-    colors = {
-        "DEBUG"    : "grey",
-        "INFO"     : "cyan",
-        "WARNING"  : "yellow",
-        "ERROR"    : "red",
-        "EXCEPTION": "magenta"
-    }
-
-    logs = None
-
-    def __init__(self, parent):
-        """
-        Parameters
-        ----------
-        parent : Tabs
-            Parent tabs object, for back-reference
-        """
-        self.parent = parent
-
-        with ui.column().classes("h-screen w-full"):
-            with ui.expansion("Settings & Stats", value=True).classes("w-full"):
-                ui.chip(
-                    "Logs are loaded on-demand as they can cause the UI to lag. It is recommended to toggle off each option when done.",
-                    icon  = "warning",
-                    color = "yellow"
-                ).props('outline square')
-
-                self.levels = ui.card().classes("border")
-
-            self.lines = ui.element().classes("nicegui-log h-full")
-
-    async def reset(self, isofit=None):
-        self.lines.clear()
-        self.levels.clear()
-
-        if isofit:
-            self.logs = isofit.logs
-
-        try:
-            self.logs.parse()
-            self.logs.filter(None)
-            self.logs.build()
-
-            self.populateLogs()
-            self.populateLevels()
-        except:
-            Logger.exception(f"Failed to load logs")
-
-            self.logs = None
-            with self.lines:
-                ui.chip(
-                    "No log file found or could not parse. If it is outside of the directory, place it somewhere inside with the extension .log",
-                    icon  = "error",
-                    color = "red"
-                ).classes("w-full")
-
-    def toggle(self, event):
-        self.logs.toggle(event.sender.text, event.value)
-        self.logs.build()
-        self.populateLogs()
-
-    def populateLevels(self):
-        self.levels.clear()
-        with self.levels:
-            for key, value in self.logs.format.items():
-                ui.switch(key, value=value, on_change=self.toggle)
-            for level, value in self.logs.selected.items():
-                ui.switch(level, value=value, on_change=self.toggle)
-
-    def populateLogs(self):
-        self.lines.clear()
-        with self.lines:
-            for (ts, lvl, msg) in self.logs.lines:
-                # color = self.logs.colors[lvl]
-                color = self.colors.get(lvl.strip(), "white")
-                with ui.row():
-                    ui.label(ts).classes("text-orange")
-                    ui.label(lvl).classes(f"text-{color}")
-                    ui.label(msg)
-
-
-class Spectra:
-    plot = None
-
-    def __init__(self, parent):
-        """
-        Parameters
-        ----------
-        parent : Tabs
-            Parent tabs object, for back-reference
-        """
-        self.parent = parent
-
-        self.active = None # Active row used for image plotting
-        self.inputs = [] # Tracks input rows
-        self.traces = [] # Tracks annotations
-        self.spectras = [] # SpectraPlot objects
-
-        self.files = observables.ObservableList([], on_change=self.setOptions)
-
-        with ui.splitter(value=30).classes("w-full h-full") as splitter:
-            with splitter.before:
-                with ui.card().classes("w-full") as self.header:
-                    with ui.button_group().classes("w-full").props("outline"):
-
-                        ui.button("Reset", on_click=self.resetImage) \
-                            .classes("flex-1") \
-                            .props("outline") \
-                            .tooltip("Removes all annotations and spectra plots")
-
-                        ui.button("Add File", on_click=self.createFileRow) \
-                            .classes("flex-1") \
-                            .props("outline") \
-                            .tooltip("This must be the same X/Y dimensions as the first file")
-
-                        with (ui.dropdown_button("Settings")
-                            .classes("flex-1")
-                            .props("outline")
-                        ):
-                            with ui.row().classes("p-4"):
-                                with ui.column():
-                                    self.r = ui.number(label="Red Band", value=60, on_change=self.createImage, precision=0).props("dense")
-                                    self.g = ui.number(label="Green Band", value=40, on_change=self.createImage, precision=0).props("dense")
-                                    self.b = ui.number(label="Blue Band", value=30, on_change=self.createImage, precision=0).props("dense")
-                                with ui.column():
-                                    self.brighten = ui.switch("Brighten Image", value=True, on_change=self.createImage)
-                                    self.annoDelClick = ui.switch("Delete via Click", value=False).tooltip("Clicking on annotation points on the image will remove the spectra plot. Alternatively, use the dropdown menu on the top-left corner of the plot")
-                                    # self.shareX = ui.switch("Share X axis", value=False)
-                                    # self.sharey = ui.switch("Share Y axis", value=False)
-
-                    self.createFileRow(main=True)
-
-                with ui.element().classes("relative w-full") as self.column:
-                    # Loading overlay
-                    with ui.card().classes("absolute inset-0 z-10 flex items-center justify-center bg-white/70") as self.loading:
-                        ui.spinner(size='xl')
-                    self.loading.visible = False
-
-                    # Plotly figure
-                    self.plot = ui.plotly(go.Figure()).classes('w-full h-full')
-
-                # Override the event to delete the source string which crashes the server when the image is too large
-                self.plot.on('plotly_click', js_handler="""
-                    (event) => {
-                        event.points.forEach(point => delete point.data.source);
-                        return emitEvent('clickedPoint', event);
-                    };
-                """)
-                ui.on('clickedPoint', self.annotationEvent)
-            with splitter.after:
-                with ui.card().classes("h-full w-full p-0"):
-                    self.scroll = ui.scroll_area().classes("h-full w-full")
-
-    async def setOptions(self, *_):
-        """
-        Updates each file selection dropdown with the self.files list
-        This is called each time the list is updated
-        """
-        for row in self.inputs:
-            row["select"].set_options(self.files)
-
-    def createFileRow(self, main=False):
-        """
-        Creates a new input row
-
-        Parameters
-        ----------
-        main : bool, default=False
-            The main row (self.inputs[0]) cannot be removed, so the delete button is
-            replaced with an info tooltip
-        """
-        row = {}
-        self.inputs.append(row)
-        color = plotlyColor(len(self.inputs)-1)
-
-        with self.header:
-            with ui.row().classes("w-full"):
-                row["colorBtn"] = ui.button(
-                    icon = "show_chart",
-                    on_click = lambda e: self.setActive(row)
-                ).props("dense outline") \
-                .classes(f"h-full") \
-                .style(f"--q-primary: {color};") \
-                .tooltip("Set this data as the image")
-
-                row["select"] = ui.select(
-                    options = self.files,
-                    label = "Select data file",
-                    on_change = lambda e: self.loadFile(e.value, row),
-                    new_value_mode = "add-unique",
-                ).classes("flex-1").props("dense")
-
-                if main:
-                    self.active = row
-                    row["colorBtn"].icon = "image"
-
-                    with (ui.button(icon="question_mark")
-                        .classes("h-full")
-                        .props("dense outline")
-                        .style(f"--q-primary: {color};")
-                    ):
-                        with ui.tooltip():
-                            ui.label("Click on the button to the left of each file input to select that data to be the image plot")
-                            ui.label("Changes to the settings will use this data")
-                            ui.label("Additional files must be the same X/Y/Wavelength shape to plot correctly. There are no enforcement of this, however")
-                            ui.label("Some plotting features are only available if certain requirements are met:")
-                            ui.label("  - Subtraction, % Difference : Only two data files are loaded")
-                            ui.label("  - Average : Two or more data files are loaded")
-                else:
-                    row["removeBtn"] = ui.button(
-                        icon="close",
-                        on_click=lambda e: self.delInput(row)
-                    ).classes("h-full") \
-                    .props("dense outline") \
-                    .style(f"--q-primary: {color};") \
-                    .tooltip("Remove this data")
-
-    async def delInput(self, row):
-        """
-        Removes an input row and its data from the spectra plots
-
-        Parameters
-        ----------
-        row : dict
-            Input row to remove
-        """
-        self.loading.visible = True
-
-        i = self.inputs.index(row)
-        self.header.remove(i+1)
-        self.inputs.remove(row)
-
-        # Update colors of buttons
-        for j, row in enumerate(self.inputs[i:]):
-            color = plotlyColor(i+j)
-            row["colorBtn"].style(f"--q-primary: {color};")
-            row["removeBtn"].style(f"--q-primary: {color};")
-
-        # Update colors of spectra plots
-        for spectra in self.spectras:
-            await spectra.build()
-
-        await self.setActive(self.inputs[0])
-
-        self.loading.visible = False
-
-    async def reset(self, *_):
-        """
-        Resets the tab
-        """
-        self.scroll.clear()
-        self.traces.clear()
-        self.spectras.clear()
-
-        self.files.clear()
-        self.files += await run.io_bound(WD.find, "rfl", all=True, exc=["hdr", "subs"])
-        self.files.sort()
-
-        if self.files:
-            self.inputs[0]["select"].set_value(self.files[0])
-
-    async def resetImage(self):
-        """
-        Resets the image, deleting all annotations and spectra plots
-        """
-        self.scroll.clear()
-        self.traces.clear()
-        self.spectras.clear()
-
-        await self.createImage()
-
-    def load(self, path):
-        """
-        TODO
-
-        Parameters
-        ----------
-        path : str
-            Path to data file to load. If the path exists, loads directly; otherwise
-            use IsofitWD to find and load
-        """
-        Logger.debug(f"Loading {path}")
-        if Path(path).exists():
-            try:
-                return Loaders.envi(path)
-            except:
-                Logger.exception("Failed to load envi")
-        else:
-            try:
-                return WD.load(path=path)
-            except:
-                Logger.exception("Failed to load via WD")
-
-    async def loadFile(self, file, row):
-        """
-        Loads a data file
-
-        Parameters
-        ----------
-        file : str
-            File to load
-        row : dict
-            Input dict associated with this file
-        """
-        self.loading.visible = True
-
-        row["file"] = file
-        row["data"] = await run.io_bound(self.load, path=file)
-
-        if row["data"] is None:
-            Logger.error("No data available, returning")
-            self.loading.visible = False
-            return
-
-        if row == self.active:
-            self.scroll.clear()
-            self.traces.clear()
-            self.spectras.clear()
-
-            # Set the limits to the band number inputs
-            min = int(row["data"].band.min())
-            max = int(row["data"].band.max())
-            for opt in {self.r, self.g, self.b}:
-                opt.min = min
-                opt.max = max
-
-            await self.createImage()
-        else:
-            for spectra in self.spectras:
-                await spectra.build()
-
-            self.loading.visible = False
-
-    async def setActive(self, row):
-        """
-        Sets the active row which is used for plotting the image
-
-        Parameters
-        ----------
-        row : dict
-            Input row to set as the active row
-        """
-        self.active = row
-
-        # Update icons
-        row["colorBtn"].icon = "image"
-        for data in self.inputs:
-            if data == row:
-                data["colorBtn"].icon = "image"
-            else:
-                data["colorBtn"].icon = "show_chart"
-
-        await self.createImage()
-
-    def buildImage(self):
-        """
-        """
-        fig = go.Figure()
-
-        if self.active:
-            try:
-                bands = [self.r.value, self.g.value, self.b.value]
-                rgb = self.active["data"].sel(band=bands).transpose("y", "x", "band")
-
-                if self.brighten.value:
-                    rgb /= rgb.max(["x", "y"])
-
-                # Convert to pixel coords for easier pixel selection
-                rgb["x"] = range(rgb.x.size)
-                rgb["y"] = range(rgb.y.size)
-
-                fig = px.imshow(rgb, template="plotly_dark")
-            except:
-                Logger.exception(f"Failed to plot spectra image")
-
-        for trace in self.traces:
-            fig.add_trace(trace)
-
-        fig.update_layout(
-            margin = dict(l=0, r=20, t=0, b=0),
-            showlegend = False,
-            paper_bgcolor = "rgba(0, 0, 0, 0)",
-        )
-
-        return fig
-
-    async def createImage(self):
-        """
-        Creates the data for the image and passes it along to the frontend
-
-        This image is always self.inputs[0], ie. the first input
-        """
-        self.loading.visible = True
-
-        self.img = await run.io_bound(self.buildImage)
-        self.plot.update_figure(self.img)
-
-        self.loading.visible = False
-
-    def setImage(self, figure):
-        """
-        Sets the frontend image's figure
-
-        Parameters
-        ----------
-        figure : go.Figure
-
-        Returns
-        -------
-
-        """
-        self.img = figure
-        figure.update_layout(
-            margin = dict(l=0, r=20, t=0, b=0),
-            paper_bgcolor = "rgba(0, 0, 0, 0)",
-            showlegend    = False
-        )
-        if self.plot:
-            self.plot.update_figure(figure)
-
-    async def plotImage(self, image):
-        """
-        Creates image figure to pass to the frontend
-
-        Parameters
-        ----------
-        image : xarray.Dataset
-            Image to plot
-        """
-        try:
-            img = px.imshow(image, template="plotly_dark")
-        except:
-            Logger.exception(f"Failed to plot spectra image")
-            img = go.Figure()
-
-        self.setImage(img)
-
-    async def addAnotation(self, x, y):
-        """
-        Adds an annotation point to the image and creates a spectra plot
-
-        Parameters
-        ----------
-        x : int
-            X coordinate
-        y : int
-            Y coordinate
-        """
-        i = len(self.img.data)
-
-        trace = go.Scatter(
-            x = [x],
-            y = [y],
-            mode = "markers+text",
-            name = f"Spectra {i} @ ({x}, {y})",
-            text = [str(i)],
-            marker = {"symbol": "circle-open"},
-            textposition = "top center",
-        )
-        self.traces.append(trace)
-        self.img.add_trace(trace)
-        await self.plotSpectra(i, x, y)
-
-        self.plot.update()
-
-    async def delAnnotation(self, i):
-        """
-        Deletes an annotation from the image as well as removes the spectra plot
-        associated with that point
-
-        Parameters
-        ----------
-        i : int
-            Index of the annotation to remove
-        """
-        # Remove the annotation from the image
-        data = list(self.img.data)
-        data.pop(i)
-        self.img.data = data
-
-        # Update names and texts
-        for trace in self.img.data[i:]:
-            pos = int(trace.text[0])
-            new = str(pos-1)
-            trace.name = re.sub(r"(\d+) @", f"{new} @", trace.name)
-            trace.text = [new]
-
-        # Convert to 0-indexed
-        i -= 1
-
-        # Remove the corresponding spectra plot
-        self.scroll.remove(i)
-        self.spectras.pop(i)
-        self.traces.pop(i)
-
-        # Update the colors
-        for spectra in self.spectras[i:]:
-            spectra.updateID(spectra.id - 1)
-
-        self.plot.update()
-
-    async def annotationEvent(self, event):
-        """
-        Event handler when an image pixel is clicked. If the pixel has an annotation,
-        it deletes it. If it doesn't, it adds an annotation.
-
-        Parameters
-        ----------
-        event : nicegui.events.GenericEventArguments
-            Event triggered when the image is clicked
-        """
-        # Images will only ever have 1 point returned
-        point = event.args["points"][0]
-        data = point["data"]
-
-        # Clicked a pixel on the image
-        if data["type"] == "image":
-            y, x = point["pointIndex"]
-            await self.addAnotation(x, y)
-
-        # Clicked an annotation
-        elif self.annoDelClick.value and data["type"] == "scatter":
-            i = point["curveNumber"]
-            await self.delAnnotation(i)
-
-    async def plotSpectra(self, i, x, y):
-        """
-        Creates a spectra plot for a pixel
-
-        Parameters
-        ----------
-        i : int
-            Index ID to assign
-        x : int
-            X coordinate
-        y : int
-            Y coordinate
-        """
-        with self.scroll:
-            spectra = SpectraPlot(i, x, y, self)
-            await spectra.build()
-
-            self.spectras.append(spectra)
-
-
-class SpectraPlot:
-    def __init__(self, id, x, y, parent):
-        """
-        Plots one or more spectra lines for a given pixel
-
-        Parameters
-        ----------
-        id : int
-            ID of the plot
-        x : int
-            X coordinate
-        y : int
-            Y coordinate
-        parent : Spectra
-            Parent object to refer to for things like close plot or data retrieval
-        """
-        self.id = id
-        self.x = x
-        self.y = y
-        self.parent = parent
-
-        self.cache = []
-
-        with ui.card().classes("relative w-full p-2"):
-            with ui.card().classes("absolute inset-0 z-20 flex items-center justify-center bg-white/70") as self.loading:
-                ui.spinner(size='xl')
-
-            # Dropdown button in top-left corner
-            with ui.button(icon="more_horiz", on_click=self.updateOptions).classes("absolute top-0 left-0 z-10").props("outline dense"):
-                with ui.menu() as menu:
-                    ui.menu_item("Close plot", on_click=self.close)
-                    self.trim = ui.switch("Remove Minimums", value=True, on_change=self.build)
-                    self.styles = ui.radio(
-                        options=["Separate", "Average", "Subtract", "% Difference"],
-                        value="Separate",
-                        on_change=self.styleChanged
-                    ).classes('flex flex-col p-1')#.props("disable")
-
-            # Plotly chart
-            blank = go.Figure(go.Scatter())
-            self.plot = ui.plotly(blank).classes("w-full")
-
-    def updateID(self, id=None):
-        """
-        Updates the plot title with the color assigned per the ID
-
-        Parameters
-        ----------
-        id : int, default=None
-            Optionally set the ID
-        """
-        if id is not None:
-            self.id = id
-
-        self.plot.figure.update_layout(
-            title={
-                "text": f"Spectra {self.id} @ ({self.x}, {self.y})",
-                "font": {
-                    "color": plotlyColor(self.id)
-                }
-            }
-        )
-
-        self.plot.update()
-
-    async def updateOptions(self):
-        """
-        Enables/disables radio options depending on the number of input files available
-
-        This is called every time the dropdown is opened so that it is always up to
-        date
-        """
-        script = f'''
-        const radioGroup = document.getElementById("c{self.styles.id}");
-        const radios = radioGroup.querySelectorAll('[role="radio"]');
-        '''
-        for i, style in enumerate(self.styles.options):
-            if self.isStyleValid(style):
-                script += f'''
-                radios[{i}].setAttribute('aria-disabled', 'false');
-                radios[{i}].classList.remove('disabled');
-                '''
-            else:
-                script += f'''
-                radios[{i}].setAttribute('aria-disabled', 'true');
-                radios[{i}].classList.add('disabled');
-                '''
-
-        ui.run_javascript(script)
-
-    async def close(self):
-        """
-        Closes this plot
-        """
-        await self.parent.delAnnotation(self.id)
-
-    async def build(self):
-        """
-        Retrieves the loaded data from the parent, selects this spectra's pixel, and
-        prepares the data for plotting
-        """
-        self.cache.clear()
-
-        for row in self.parent.inputs:
-            if (data := row.get("data")) is not None:
-                try:
-                    spectra = data.isel(x=self.x, y=self.y)
-                    spectra = spectra.rename(wavelength='Wavelength')
-
-                    # Remove minimum values
-                    if self.trim.value:
-                        spectra = spectra.where(spectra != spectra.min())
-
-                    spectra.name = "Reflectance"
-                    df = spectra.to_dataframe()
-
-                    # Only cache these columns (sometimes fwhm tags along)
-                    df = df[["Wavelength", "Reflectance"]]
-
-                    self.cache.append(df)
-                except:
-                    Logger.exception(f"Failed to create a spectra plot for {row['file']}")
-
-        await self.draw()
-
-    def calcStyle(self):
-        """
-        Performs the calculations for the style options available
-
-        Returns
-        -------
-        list
-            The list of dataframes prepared for plotting
-        """
-        style = self.styles.value
-
-        if style == "Separate":
-            return self.cache
-
-        calc = self.cache[0].copy()
-        rfls = [df["Reflectance"] for df in self.cache]
-
-        if style == "Subtract":
-            calc["Reflectance"] = rfls[0] - rfls[1]
-
-        elif style == "% Difference":
-            a = rfls[0]
-            b = rfls[1]
-
-            calc["Reflectance"] = (a - b).abs() / ((a + b) / 2) * 100
-
-        elif style == "Average":
-            calc["Reflectance"] = sum(rfls) / len(rfls)
-
-        return [calc]
-
-    async def draw(self):
-        """
-        Plots the data with the set style
-        """
-        self.loading.visible = True
-
-        data = self.calcStyle()
-
-        fig = go.Figure()
-        for i, df in enumerate(data):
-            fig.add_trace(
-                trace = px.line(df,
-                    x = 'Wavelength',
-                    y = 'Reflectance',
-                ).data[0]
-            )
-
-        for i, trace in enumerate(fig.data):
-            trace.line.color = plotlyColor(i)
-
-        fig.update_layout(
-            height = 300,
-            margin = dict(l=0, r=0, t=30, b=0),
-            template = "plotly_dark",
-            paper_bgcolor = "rgba(0, 0, 0, 0)"
-        )
-
-        self.plot.update_figure(fig)
-
-        self.updateID()
-
-        self.loading.visible = False
-
-    def isStyleValid(self, style):
-        """
-        Checks if a selected style is valid given the data available
-
-        Parameters
-        ----------
-        style : str
-            Style mode to check validity
-        """
-        if style != "Separate":
-            count = len(self.cache)
-            if count < 2:
-                return False
-            if style in {"Subtract", "% Difference"} and count != 2:
-                return False
-
-        return True
-
-    async def styleChanged(self, style):
-        """
-        Intercepts the radio style change to check if the style is a valid option
-        If not, discards the change
-
-        Parameters
-        ----------
-        style : nicegui.events.ValueChangeEventArguments
-            Event triggered by the `styles` ui.radio
-        """
-        if not self.isStyleValid(style.value):
-            return self.styles.set_value("Separate")
-        await self.draw()
-
-
-class Configs:
-    data = None
-
-    def __init__(self, parent):
-        """
-        Parameters
-        ----------
-        parent : Tabs
-            Parent tabs object, for back-reference
-        """
-        self.parent = parent
-
-        with ui.row():
-            self.select = ui.select([], label="Select a Configuration", on_change=self.loadConfig)
-            ui.switch("Editable", value=False, on_change=self.readOnly).classes("h-full")
-
-        self.editor = ui.json_editor({'content': {'json': {}}, 'readOnly': True}).classes('w-full jse-theme-dark')
-
-    async def reset(self, isofit=None):
-        configs = WD.find("config/.json", all=True)
-        self.select.set_options(configs, value=configs[0])
-        # if isofit:
-        #     self.isofit = isofit
-        #     configs = isofit.find("config/.json", all=True)
-        #     self.select.set_options(configs, value=configs[0])
-
-    def loadConfig(self, event):
-        # data = self.isofit.load(path=event.value)
-        data = WD.load(path=event.value)
-        self.editor.run_editor_method('updateProps', {'content': {'json': data}})
-
-    def readOnly(self, event):
-        self.editor.run_editor_method('updateProps', {'readOnly': not event.value})
-
-
-#%%
-
-def multiplot(figs=[], height=300):
-    """
-    Creates a multi plot object that shares X and Y axes
-
-    Parameters
-    ----------
-    figs : list[go.Figure]
-        List of subplots to add
-    height : int, default=300
-        Height of each subplot
-
-    Returns
-    -------
-    fig : go.Figure
-        Plotly figure with one or more subplots
-    """
-    # Ensure at least one figure is set
-    if not figs:
-        figs = [go.Figure()]
-
-    fig = go.Figure()
-    fig.set_subplots(
-        rows = len(figs),
-        cols = 1,
-        shared_xaxes = 'all',
-        shared_yaxes = 'all',
-        vertical_spacing = 0.01
-    )
-    fig.update_layout(**{
-        "margin": dict(l=0, r=20, t=0, b=0),
-        "paper_bgcolor": "rgba(0, 0, 0, 0)",
-        "template": "plotly_dark",
-        "height": height*len(figs),
-    })
-
-    for i, plot in enumerate(figs):
-        for trace in plot.data:
-            # fig.update_layout(legendgroup=f"{i+1}")
-            fig.add_trace(trace, row=i+1, col=1)
-
-    # TODO: Get subplots to have their own legends
-    # Does not work
-    try:
-        for i, yaxis in enumerate(fig.select_yaxes()):
-            i += 1
-            legend_name = f"legend{i}"
-            if yaxis.domain is not None:
-                fig.update_layout({legend_name: dict(y=yaxis.domain[1], yanchor="top")}, showlegend=True)
-                fig.update_traces(row=i, legend=legend_name)
-    except:
-        pass
-
-    return fig
-
-#%%
-
-class MultiPlotLUT:
-    """
-    Handles the construction and management of a single LUT plotting card in the LUTs
-    tab
-    """
-    def __init__(self, parent, files=None, cache=None):
-        """
-        Parameters
-        ----------
-        parent : LUT
-            Parent LUT class that hosts this object
-        files : list, default=None
-            Shared NiceGUI observables for the files options, otherwise will be an
-            independent list
-        cache : dict, default=None
-            Shared cache dict if provided, otherwise creates one for this object
-        """
-        self.parent = parent
-
-        if files is None:
-            files = observables.ObservableList([], on_change=self.setOptions)
-        self.files = files
-
-        if cache is None:
-            cache = {}
-        self.cache = cache
-
-        self.main = self.new
-        self.plots = [self.main]
-
-        # Defaults
-        self.quant = None
-        self.dim = "None"
-
-        with ui.card().classes("w-full border"):
-            with ui.column().classes("w-full"):
-                with ui.row().classes("w-full items-center") as self.header:
-                    self.main["select"] = ui.select(
-                        label = "LUT File",
-                        options = self.files,
-                        multiple = False, # TODO?
-                        new_value_mode = "add-unique",
-                        on_change = lambda e: self.changeFile(e.value)
-                    ).classes("w-[56%]").props('use-chips')
-
-                    self.quants = ui.select(
-                        label = "Select a LUT quantity",
-                        options = [],
-                        on_change = lambda e: self.changeQuant(e.value)
-                    ).classes("w-[15%]")
-                    self.quants.disable()
-
-                    self.dims = ui.select(
-                        label = "Select a LUT dimension",
-                        options = [],
-                        on_change = lambda e: self.changeDim(e.value)
-                    ).classes("w-[15%]")
-                    self.dims.disable()
-
-                    ui.button(
-                        icon = "question_mark",
-                        # on_click = lambda: self.deleteSubplot(plot),
-                    ).props("outline round") \
-                    .classes("ml-auto") \
-                    .tooltip("Select a LUT file or provide a path to a LUT file to load the quantities. Then select a quantity to load the dimensions. Finally, select a dimension to plot.")
-
-                    ui.button(
-                        icon = "add",
-                        on_click = self.createSubplot,
-                    ).props("outline round") \
-                    .classes("ml-auto") \
-                    .tooltip("Add subplot. This will share X and Y axes as well as quantity and dimension with the main plot")
-
-                    ui.button(
-                        icon = "close",
-                        on_click = lambda: parent.deletePlot(self)
-                    ).props("outline round") \
-                    .classes("ml-auto") \
-                    .tooltip("Removes this entire plot group")
-
-                # Relative wrapper to control layering
-                with ui.element().classes("relative w-full") as self.column:
-                    # Loading overlay
-                    with ui.card().classes("absolute inset-0 z-10 flex items-center justify-center bg-white/70") as self.loading:
-                        ui.spinner(size='xl')
-                    self.loading.visible = False
-
-                    # Plotly figure
-                    self.ui = ui.plotly(multiplot()).classes("w-full")
-
-    @property
-    def new(self):
-        """
-        Creates a dict with default values to be used for a new plot
-
-        Returns
-        -------
-        dict
-            Default values for a new plot
-        """
-        return {
-            "plot": self.blank,
-            "lut": None,
-            "select": None,
-        }
-
-    @property
-    def blank(self):
-        """
-        Creates a blank plotly figure with an empty scatter trace so that it renders
-        the plot in the GUI. Without the scatter, the plot will be invisible
-
-        Returns
-        -------
-        go.Figure
-            Blank plotly figure
-        """
-        return go.Figure(go.Scatter())
-
-    def load(self, file):
-        """
-        Loads a LUT dataset and stores it in the cache
-
-        Parameters
-        ----------
-        file : str
-            Path to a LUT file to load. If the path does not exist, will attempt to
-            find it under the WD
-
-        Returns
-        -------
-        xr.Dataset
-            Loaded LUT dataset
-        """
-        self.loading.visible = True
-
-        Logger.debug(f"Loading path: {file}")
-        if file not in self.cache:
-            if Path(file).exists():
-                try:
-                    self.cache[file] = luts.load(file).unstack()
-                except Exception as e:
-                    Logger.exception(f"Failed to load via luts.py")
-            else:
-                try:
-                    self.cache[file] = WD.load(path=file).unstack()
-                except Exception as e:
-                    Logger.exception(f"Failed to load via WD")
-
-        self.loading.visible = False
-
-        return self.cache.get(file)
-
-    def setOptions(self, *_):
-        """
-        Updates each file selection dropdown with the self.files list
-        This is called each time the list is updated
-        """
-        for plot in self.plots:
-            plot["select"].set_options(self.files)
-
-    async def createSubplot(self):
-        """
-        Adds a new subplot to the figure
-        """
-        plot = self.new
-        self.plots.append(plot)
-
-        with self.header:
-            with ui.row().classes("w-full items-center"):
-                plot["select"] = ui.select(
-                    label = f"LUT File for Plot {len(self.plots)}",
-                    options = self.files,
-                    new_value_mode = "add-unique",
-                    on_change = lambda e: self.updateSubplot(plot, file=e.value)
-                ).classes("w-[96%]")
-
-                ui.button(
-                    icon = "close",
-                    on_click = lambda: self.deleteSubplot(plot),
-                ).props("outline round").tooltip("Remove subplot").classes("ml-auto")
-
-        # Shift to bottom of the card
-        # self.ui.move(target_index=-1)
-
-        # Update the figure with a blank
-        await self.updateUI()
-
-    async def updateSubplot(self, plot, file):
-        """
-        Updates a subplot with a new LUT input
-
-        Parameters
-        ----------
-        plot : dict
-            Plot data dict
-        file : str
-            LUT file to load
-        """
-        lut = await run.io_bound(self.load, file)
-        if lut is None:
-            return
-
-        plot["lut"] = lut
-        plot["plot"] = await self.plot(
-            lut   = plot["lut"],
-            quant = self.quant,
-            dim   = self.dim
-        )
-        await self.updateUI()
-
-    async def deleteSubplot(self, plot):
-        """
-        Deletes a subplot and removes it from the UI
-
-        Parameters
-        ----------
-        plot : dict
-            Plot data dict
-        """
-        i = self.plots.index(plot)
-        self.header.remove(i+5) # 5 to skip the primary header components
-        self.plots.pop(i)
-
-        await self.updateUI()
-
-        # Update labels for consistency
-        for i, plot in enumerate(self.plots[1:]):
-            plot["select"].props(f'label="LUT File for Plot {i+2}"')
-
-    async def changeFile(self, file):
-        """
-        Changes the primary LUT dataset and updates the quantities list
-
-        Parameters
-        ----------
-        file : str
-            Path to LUT file to load
-        """
-        self.quants.disable()
-        self.dims.disable()
-
-        lut = await run.io_bound(self.load, file)
-        if lut is None:
-            return
-
-        self.main["lut"] = lut
-
-        # Update the quantities list
-        plottable = set(lut) - set(lut.drop_dims("wl"))
-        if plottable:
-            self.quants.set_options(sorted(plottable))
-            self.quants.enable()
-
-        # Reset the quant selection
-        self.quants.set_value(None)
-
-    async def changeQuant(self, quant):
-        """
-        Changes the selected quantity from the active main LUT file and updates the
-        dimensions list
-
-        Parameters
-        ----------
-        quant : str
-            Quantity of the LUT to plot
-        """
-        self.quant = quant
-
-        lut = self.main["lut"]
-
-        # Discover the available dims
-        dims = []
-        if quant in lut:
-            dims = set(lut[quant].dims) - {"wl",}
-
-        # Set them as the options, including "None" to average over all dims
-        self.dims.set_options(["None"] + sorted(dims))
-
-        if dims:
-            self.dims.enable()
-        else:
-            self.dims.disable()
-
-        # Attempt to re-plot with the last chosen dim
-        if self.dim != self.dims.value:
-            self.dims.set_value(self.dim)
-        else:
-            await self.changeDim(self.dim)
-
-    async def changeDim(self, dim):
-        """
-        Changes the plotting dimension
-
-        Parameters
-        ----------
-        dim : str
-            Dimension to plot along. All other dimensions will be squeezed to the mean
-        """
-        self.dim = dim
-
-        # Update all plots
-        for plot in self.plots:
-            plot["plot"] = await self.plot(
-                lut   = plot["lut"],
-                quant = self.quant,
-                dim   = self.dim
-            )
-
-        await self.updateUI()
-
-    async def plot(self, lut, quant, dim):
-        """
-        Attempts to create a LUT plot. If it fails, returns a blank figure
-
-        Parameters
-        ----------
-        lut : xr.Dataset
-            LUT data object
-        quant : str
-            Quantity to select
-        dim : str
-            Dimension to plot along. All other dimensions will be averaged
-
-        Returns
-        -------
-        go.Figure
-            Either the plotted figure or a blank if it failed
-        """
-        self.loading.visible = True
-
-        try:
-            data = lut[quant]
-            dims = set(data.coords) - {"wl", dim}
-
-            if dims:
-                mean = data.mean(dims, skipna=True)
-                data = mean.rename(wl="Wavelength")
-
-                # Split each value of the dim into its own variable
-                if dim != "None":
-                    data = data.to_dataset(dim)
-
-            df = data.to_dataframe()
-
-            return px.line(df,
-                template = "plotly_dark",
-                labels = {"variable": str(dim)}
-            )
-        except:
-            Logger.exception(f"Failed to plot {quant}[{dim}]")
-            return self.blank
-        finally:
-            self.loading.visible = False
-
-    async def updateUI(self):
-        """
-        Updates the GUI with a newly constructed multi-plot figure
-        """
-        plots = [plot["plot"] for plot in self.plots]
-        fig = multiplot(plots)
-        self.ui.update_figure(fig)
-
-
-class LUTs:
-    def __init__(self, parent):
-        """
-        Parameters
-        ----------
-        parent : Tabs
-            Parent tabs object, for back-reference
-        """
-        self.parent = parent
-
-        # Stores MultiPlotLUT objects
-        self.plots = []
-
-        # Shared LUT cache
-        self.cache = {}
-
-        # Shared file options
-        self.files = observables.ObservableList([], on_change=self.setOptions)
-
-        with ui.scroll_area().classes("h-full w-full") as self.scroll:
-            self.addPlotBtn = ui.button(
-                "Add New Independent Plot",
-                on_click = self.asyncCreatePlot
-            ).classes("w-full") \
-            .props("outline")
-
-        # Create an initial plot
-        self.createPlot()
-
-    async def setOptions(self, *_):
-        """
-        Calls each MultiPlotLUT to set the file options
-
-        This function is called any time the self.files list is updated, such as when
-        the WD is changed or a manual input is given
-        """
-        for plot in self.plots:
-            plot.setOptions()
-
-    def createPlot(self):
-        """
-        Creates a new plot group using the MultiPlotLUT class
-        """
-        with self.scroll:
-            plot = MultiPlotLUT(self, files=self.files, cache=self.cache)
-            self.plots.append(plot)
-
-        # Move the button to the bottom
-        self.addPlotBtn.move(target_index=-1)
-
-    async def asyncCreatePlot(self):
-        """
-        Async wrapper for createPlot
-        """
-        self.createPlot()
-
-    async def deletePlot(self, plot):
-        """
-        Removes a plot group from the interface permanently
-
-        Parameters
-        ----------
-        plot : MultiPlotLUT
-            Card to be removed
-        """
-        i = self.plots.index(plot)
-        self.scroll.remove(i)
-        self.plots.pop(i)
-
-    async def reset(self, *_):
-        """
-        Resets the the file options when the WD changes
-        """
-        self.files.clear()
-        self.files += await run.io_bound(WD.find, "lut/.nc", all=True)
-        self.files.sort()
-
-
-def toNiceGUITree(tree=None, *, path=None, nodes=None):
-    """
-    Recursively converts an IsofitWD tree (with info) into a NiceGUI-compatible data
-    structure for the tree component
-
-    Parameters
-    ----------
-    tree : dict, default=None
-        Tree object created by IsofitWD.getTree(info=True)
-    nodes : list, default=None
-        Converted nodes for the tree component, this likely should be left as the
-        default
-
-    Returns
-    -------
-    nodes : list
-        Converted nodes for the tree component
-    """
-    if tree is None:
-        tree = WD.getTree(info=True)
-
-    if nodes is None:
-        nodes = [{"id": "root", "desc": None, "children": []}]
-        toNiceGUITree(tree, path=".", nodes=nodes[-1]["children"])
-    else:
-        for file in tree:
-            if isinstance(file, dict):
-                for subdir, files in file.items():
-                    nodes.append({
-                        "id"      : f"{path}/{subdir.name}",
-                        "label"   : subdir.name,
-                        "desc"    : subdir.info,
-                        "children": []
-                    })
-                    toNiceGUITree(files,
-                        path  = nodes[-1]["id"],
-                        nodes = nodes[-1]["children"]
-                    )
-            else:
-                nodes.append({
-                    "id"   : f"{path}/{file.name}",
-                    "label": file.name,
-                    "desc" : file.info
-                })
-
-    return nodes
-
-
-class EnhancedTree(ui.tree):
-    @property
-    def root(self):
-        return self.props["nodes"]
-
-    @property
-    def branches(self):
-        """
-        Returns branch nodes, no leaves
-
-        Yields
-        ------
-        branch : dict
-            A branch node
-        """
-        for branch in self.root:
-            if "children" in branch:
-                yield branch
-
-    def traverse(self, func, id):
-        """
-        Traverses the branches applying the given function until a value is found
-
-        Parameters
-        ----------
-        func : function
-            Function to call
-        id : str
-            ID of the node to find
-
-        Returns
-        -------
-        any
-            If the called function returns a value, returns that value
-        """
-        for branch in self.branches:
-            if value := func(id, branch=branch):
-                return value
-
-    def getPath(self, id, *, branch=None):
-        """
-        Recursively discover the string path to a given node ID
-
-        Parameters
-        ----------
-        id : str
-            ID of the node to find
-        branch : dict
-            Branch of the tree being worked on
-
-        Returns
-        -------
-        str
-            String path, eg. "configs/topoflux/config.json"
-        """
-        if branch is None:
-            return self.traverse(self.getPath, id)
-
-        for node in branch["children"]:
-            if node["id"] == id:
-                return id
-            elif "children" in node:
-                if path := self.getPath(id, branch=node):
-                    return f"{node['id']}/{path}"
-
-    def findNode(self, id, *, branch=None):
-        """
-        Find the source node for a given ID
-
-        Parameters
-        ----------
-        id : str
-            ID of the node to find
-        branch : dict
-            Branch of the tree being worked on
-
-        Returns
-        -------
-        str
-            String path, eg. "configs/topoflux/config.json"
-        """
-        if branch is None:
-            return self.traverse(self.findNode, id)
-
-        for node in branch["children"]:
-            if node["id"] == id:
-                return node
-            elif "children" in node:
-                node = self.findNode(id, branch=node)
-                if node:
-                    return node
-
-    def findSiblings(self, id, *, branch=None):
-        """
-        Finds the IDs of the sibling nodes in the tree
-
-        Parameters
-        ----------
-        id : str
-            ID of the node to find
-        branch : dict
-            Branch of the tree being worked on
-
-        Returns
-        -------
-        str
-            String path, eg. "configs/topoflux/config.json"
-        """
-        if branch is None:
-            return self.traverse(self.findSiblings, id)
-
-        # Get the sibling directories of the source node
-        nodes = branch["children"]
-        siblings = [node["id"] for node in nodes if "children" in node]
-
-        if id in siblings:
-            return siblings
-
-        # Try to find it in some child
-        for node in nodes:
-            if "children" in node:
-                if siblings := self.findSiblings(id, branch=node):
-                    return siblings
-
-
-class Setup:
-    def __init__(self, parent):
-        """
-        Parameters
-        ----------
-        parent : Tabs
-            Parent tabs object, for back-reference
-        """
-        self.parent = parent
-
-        self.pathBtns = ui.button_group().props("outline").classes("w-full")
-
-        with ui.row().classes("w-full items-center"):
-            with ui.button("Set as Working Directory",
-                on_click = self.setWD
-            ).props("outline"):
-                with ui.tooltip():
-                    ui.label("Set this path as the active working directory")
-                    ui.label("This will recursively search for ISOFIT products")
-                    ui.label("Only do this on known directories as this is an expensive operation")
-                    ui.label("Executing this on protected or excessively deep directories may crash this program's server")
-
-            self.search = ui.input("Search",
-                on_change = self.glob
-            ).classes("flex-grow")
-
-        self.preview = ui.column().classes("w-full")
-
-        self.filter = ui.input("filter").classes("w-full")
-        self.filter.visible = False
-        self.directoryTree = ui.scroll_area().classes("w-full h-full")
-        self.stepper()
-
-    def addPathButtons(self):
-        """
-        """
-        setWD = lambda e: self.search.set_value(Config["Paths"][e.sender.text])
-
-        self.pathBtns.clear()
-        if "Paths" in Config:
-            with self.pathBtns:
-                for name, path in Config["Paths"].items():
-                    if not Path(path).exists():
-                        Logger.error(f"Path not found: {path}")
-                        continue
-
-                    ui.button(name, on_click=setWD).props("outline").classes("flex-grow")
-
-    async def glob(self, path):
-        """
-        Globs the filesystem at the current search path, placing directories before
-        files
-
-        Parameters
-        ----------
-        path : ValueChangeEventArguments
-            Event when the search string changes
-        """
-        self.preview.clear()
-        with self.preview:
-            ui.spinner('dots', size='xl')
-
-        path = Path(path.value)
-        if not path.exists():
-            glob = path.parent.glob(f"{path.name}*")
-        else:
-            glob = path.glob("*")
-
-        dirs = []
-        files = []
-        for file in glob:
-            if file.is_dir():
-                dirs.append(file.name)
-            else:
-                files.append(file.name)
-
-        data = [
-            {"path": ".."}
-        ] + [
-            {"path": f"{d}/"}
-            for d in sorted(dirs)
-        ] + [
-            {"path": f}
-            for f in sorted(files)
-        ]
-
-        self.preview.clear()
-        with self.preview:
-            if data:
-                ui.aggrid({
-                    "columnDefs": [
-                        {"field": "path"}
-                    ],
-                    "rowData": data
-                }, theme="balham-dark") \
-                .classes("hide-header") \
-                .on('cellClicked',
-                    lambda event: self.appendSearch(event.args["value"])
-                )
-            else:
-                ui.icon("report_problem").tooltip("An error has occurred")
-
-    def appendSearch(self, path):
-        """
-        Appends a string to the current search string
-
-        Parameters
-        ----------
-        path : str
-            Path to append
-        """
-        current = Path(self.search.value)
-        if not current.exists():
-            current = current.parent
-        new = (current / path).resolve()
-
-        self.search.set_value(str(new))
-
-    def stepper(self):
-        """
-        Leverages a NiceGUI Stepper to provide loading information
-
-        Returns
-        -------
-        stepper : ui.stepper
-            Stepper object
-        """
-        self.directoryTree.clear()
-        with self.directoryTree:
-            with ui.stepper().props("vertical").classes("w-full") as stepper:
-                with ui.step(f"Setting path"):
-                    ui.label("Select a path above to get started")
-                    ui.label("Having an active working directory simplifies many of the UI features")
-                    ui.label("Alternatively, select tabs on the left and manually input products")
-                    ui.label("While processing a path, the tabs will be disabled to protect against thread race conditions")
-                with ui.step("Recursively searching for ISOFIT products"):
-                    ui.skeleton().classes("w-full")
-                    ui.label("Depending on the depth of the chosen directory, this may take a moment")
-                with ui.step("Building directory tree"):
-                    ui.skeleton().classes("w-full")
-                    ui.label("This typically takes the longest to parse the directory tree into a readable format")
-                    ui.label("Please be patient, there is no progress indicator but it is working")
-                    ui.label("If this takes more than two minutes, please open an issue on the repository with a copy of your terminal logs")
-                with ui.step("Rendering UI"):
-                    ui.skeleton().classes("w-full")
-
-        return stepper
-
-    async def reset(self, *_):
-        """
-        """
-        self.addPathButtons()
-
-    async def setWD(self):
-        """
-        Sets the global IsofitWD object
-        """
-        # Disable buttons while loading
-        self.parent.toggleTabs()
-
-        path = self.search.value
-        self.filter.visible = False
-
-        stepper = self.stepper()
-        stepper.next() # Setting path, done
-
-        await run.io_bound(WD.reset, path, recursive=True)
-        self.parent.resetTabs()
-        stepper.next() # Recursively search, done
-
-        data = await run.io_bound(toNiceGUITree)
-        data[0]["label"] = str(WD.path)
-        stepper.next() # Building tree, done
-
-        # Makes the UI feel more responsive by giving a slight pause after the last step
-        await sleep(0.5)
-
-        # Re-enable the tabs
-        self.parent.toggleTabs()
-
-        # Set the output directory tree
-        self.directoryTree.clear()
-        with self.directoryTree:
-            if data:
-                # ui.label("Click on an output file below to jump to an interactive component for that file [Work in Progress]")
-                # ui.label("Detected paths:")
-
-                self.tree = EnhancedTree(data, on_select=lambda e: self.navToFile(e)) \
-                .classes("border h-full w-full") \
-                .props("no-transition dense")
-
-                self.tree.expand(["root"])
-                self.tree.add_slot('default-header', '''
-                    <span :props="props">{{ props.node.label }}</strong></span>
-                ''')
-                self.tree.add_slot('default-body', '''
-                    <span :props="props">{{ props.node.desc }}</span>
-                ''')
-
-                self.filter.visible = True
-                self.filter.bind_value_to(self.tree, "filter")
-            else:
-                ui.icon("report_problem").tooltip("An error has occurred")
-
-    def navToFile(self, event):
-        """
-        TODO
-
-        Parameters
-        ----------
-        event
-        """
-        name = event.value
-
-        # Find the source node clicked
-        if node := self.tree.findNode(name):
-            if "children" in node:
-                siblings = self.tree.findSiblings(name)
-
-                self.tree.collapse(siblings)
-                self.tree.expand([name])
-            else:
-                # Get the string path to the node
-                path = Path(self.tree.getPath(name))
-
-                if len(path.parents) > 1:
-                    dir = path.parents[-2]
-                    path = path.relative_to(dir)
-
-                    # self.parent.navToFile(dir, path)
-                else:
-                    Logger.error("Cannot load files on the base")
-
-
-class Tabs:
-    disabled = False
-    default = "setup"
-    reset = {}
-    tabs = {
-        "setup"  : Setup,
-        # "config" : Configs,
-        # "logs"   : Logs,
-        "spectra": Spectra,
-        "luts"   : LUTs,
-    }
-
-    def __init__(self):
-        with ui.splitter(value=5).classes("w-full h-full") as splitter:
-            with splitter.before:
-                with ui.tabs().props('vertical').classes('w-full h-full') as tabs:
-                    self.buttons = {
-                        "setup"  : ui.tab('setup',   icon='home'),
-                        # "config" : ui.tab('config',  icon='settings'),
-                        # "logs"   : ui.tab('logs',    icon='density_small'),
-                        "spectra": ui.tab('spectra', icon='show_chart'),
-                        "luts"   : ui.tab('luts',    icon='ssid_chart'),
-                    }
-            with splitter.after:
-                with ui.tab_panels(tabs, value=self.buttons[self.default], on_change=self.tabSelected).classes('w-full h-full'):
-                    for name, btn in self.buttons.items():
-                        with ui.tab_panel(btn):
-                            # Initialize the objects
-                            self.tabs[name] = self.tabs[name](self)
-
-    async def tabSelected(self, event):
-        """
-        Awaits the tab.reset task upon selection
-        """
-        await self.tabs[event.value].resetTask
-
-    def resetTabs(self):
-        """
-        Resets all tabs by cancelling any running task and restarting
-        """
-        for tab, obj in self.tabs.items():
-            if hasattr(obj, "resetTask"):
-                obj.resetTask.cancel()
-            obj.resetTask = create_task(obj.reset())
-
-    def toggleTabs(self):
-        """
-        Toggles the disabled class of the tab buttons
-        """
-        for name in self.tabs:
-            btn = self.buttons[name]
-
-            if self.disabled:
-                btn.enable()
-            else:
-                btn.disable()
-
-        self.disabled = not self.disabled
-
-
 GUI = Tabs()
 app.on_startup(GUI.resetTabs)
-
-
-def portAvail(port: int) -> bool:
-    """
-    checks if a port is available
-
-    Parameters
-    ----------
-    port : int
-        Port to check
-
-    Returns
-    -------
-    bool
-        True if available, False otherwise
-    """
-    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-        return s.connect_ex(('localhost', port)) != 0
-
-
-def checkPorts(start : int, end : int = 100) -> None:
-    """
-    Checks for an open port between the start and end range
-
-    Parameters
-    ----------
-    start : int
-        Starting port
-    end : int
-        Offset end from the starting point to check in range of
-    """
-    if not portAvail(start):
-        for port in range(start+1, start+end):
-            Logger.info(f"Checking {port}")
-            if portAvail(port):
-                Logger.warning(f"Port {start} is already in use, recommend using --port {port}")
-                break
-        else:
-            Logger.error(f"All ports between {start}-{start+end} are in use")
 
 
 def launch(path=".", config=None, check=False, **kwargs):
@@ -2020,16 +184,16 @@ def launch(path=".", config=None, check=False, **kwargs):
 
     if path:
         Logger.info(f"Setting path: {path}")
-        GUI.tabs["setup"].search.set_value(path)
+        GUI.tabs["Setup"].search.set_value(path)
 
     if config is None:
-        config = Path(__file__).parent / "paths.ini"
+        config = Path(__file__).parent / "static" / "paths.ini"
 
     if Path(config).exists():
         Logger.info(f"Reading config: {config}")
         Config.read(config)
 
-        GUI.tabs["setup"].addPathButtons()
+        GUI.tabs["Setup"].addPathButtons()
 
     Logger.info("Launching")
     ui.run(**kwargs)
@@ -2057,6 +221,3 @@ if __name__ == "__main__":
 elif __name__ == "__mp_main__":
     logging.basicConfig(level="DEBUG")
     launch()
-
-
-#%%
