--- isoplots/isonice/paths.ini
@@ -1,2 +1,0 @@
-[Paths]
-name = /example/path/to/add/to/the/gui
