+++ isoplots/isonice/static/paths.ini
@@ -1,0 +1,8 @@
+[Paths]
+name = /example/path/to/add/to/the/gui
+NEON = /Users/jamesmo/projects/isofit/research/NEON.bak/output/NIS01_20210403_173647/
+emit = /Users/jamesmo/projects/isofit/research/jemit/
+Pasadena = /Users/jamesmo/projects/isofit/extras/examples/20171108_Pasadena
+Test6 = /Users/jamesmo/Downloads/firefox/Test6
+research = /Users/jamesmo/projects/isofit/research/
+cluster = /store/brodrick/emit/20250429_priors
